refactor: update fieldType filename
