+++ libs/common/src/models/api/field.api.ts
@@ -1,4 +1,4 @@
-import { FieldType } from "../../enums/fieldType";
+import { FieldType } from "../../enums/field-type";
 import { LinkedIdType } from "../../enums/linkedIdType";
 import { BaseResponse } from "../response/base.response";
