+++ libs/common/src/vault/services/cipher.service.ts
@@ -8,7 +8,7 @@ import { LogService } from "../../abstractions/log.service";
 import { SearchService } from "../../abstractions/search.service";
 import { SettingsService } from "../../abstractions/settings.service";
 import { StateService } from "../../abstractions/state.service";
-import { FieldType } from "../../enums/fieldType";
+import { FieldType } from "../../enums/field-type";
 import { UriMatchType } from "../../enums/uriMatchType";
 import { sequentialize } from "../../misc/sequentialize";
 import { Utils } from "../../misc/utils";
