+++ libs/importer/spec/onepassword-1pif-importer.spec.ts
@@ -1,4 +1,4 @@
-import { FieldType } from "@bitwarden/common/enums/fieldType";
+import { FieldType } from "@bitwarden/common/enums/field-type";
 
 import { OnePassword1PifImporter } from "../src/importers";
