+++ libs/importer/spec/onepassword-1pux-importer.spec.ts
@@ -1,4 +1,4 @@
-import { FieldType } from "@bitwarden/common/enums/fieldType";
+import { FieldType } from "@bitwarden/common/enums/field-type";
 import { SecureNoteType } from "@bitwarden/common/enums/secureNoteType";
 import { Utils } from "@bitwarden/common/misc/utils";
 import { CipherType } from "@bitwarden/common/vault/enums/cipher-type";
